R6 class storing a pool of workers

R6 class storing a pool of workers

R6 class storing a pool of workers

Details

A WorkerPool is a container that holds one or more workers and can request them to assign tasks, start tasks, and complete tasks. It can also check to see if any worker sessions have crashed and restart them as needed. fields and methods that allow it to work on Tasks

Methods

Public methods

WorkerPool$new()

WorkerPool$get_pool_worker()

WorkerPool$get_pool_state()

WorkerPool$try_assign()

WorkerPool$try_start()

WorkerPool$try_finish()

WorkerPool$refill_pool()

WorkerPool$shutdown_pool()

WorkerPool$clone()

Method new()

Create a new worker pool

Usage

WorkerPool$new(workers = 4L)

Arguments

workers

The number of workers in the pool.

Returns

A new WorkerPool object.

Method get_pool_worker()

Return a specific worker

Usage

WorkerPool$get_pool_worker(ind)

Arguments

ind

The index of the worker in the pool

Returns

The corresponding Worker object.

Method get_pool_state()

Return a simple summary of the worker pool

Usage

WorkerPool$get_pool_state()

Returns

A named character vector specifying the current state of each worker ("starting", "idle", "busy", or "finished"). Names denote worker ids

Method try_assign()

Attempt to assign tasks to workers. This method is intended to be called by TaskQueue objects.

Usage

WorkerPool$try_assign(tasks)

Arguments

tasks

A list of Task objects

Returns

Invisibly returns NULL

Method try_start()

Attempt to start any assigned but not-yet-started tasks in the worker pool. This method is intended to be called by TaskQueue objects.

Usage

WorkerPool$try_start()

Returns

Invisibly returns NULL

Method try_finish()

Attempt to finish any started but not-yet-returned tasks in the worker pool. This method is intended to be called by TaskQueue objects.

Usage

WorkerPool$try_finish()

Returns

Invisibly returns NULL

Method refill_pool()

Check all workers in the pool looking for workers that have crashed or been shutdown, and replace them with fresh workers.

Usage

WorkerPool$refill_pool()

Returns

This function is called primarily for its side effect. It returns a named character documenting the outcome, indicating the current state of each worker: should not be "finished" for any worker. Names denote worker ids.

Method shutdown_pool()

Terminate all workers in the worker pool.

Usage

WorkerPool$shutdown_pool(grace = 1000)

Arguments

grace

Grace period in milliseconds. If a worker process is still running after this period, it will be killed.

Returns

This function is called primarily for its side effect. It returns a named character documenting the outcome, indicating the current state of each worker: should be "finished" for all workers. Names denote worker ids.

Method clone()

The objects of this class are cloneable with this method.

Usage

WorkerPool$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.