R6 class for a multi-worker task queue

R6 class for a multi-worker task queue

R6 class for a multi-worker task queue

Details

Documentation baby...

Methods

Public methods

Queue$new()

Queue$push()

Queue$run()

Queue$retrieve()

Queue$get_queue_report()

Queue$get_queue_workers()

Queue$clone()

Method new()

Create a task queue

Usage

Queue$new(workers = 4L)

Arguments

workers

Either the number of workers to employ in the task queue, or a WorkerPool object to use when deploying the tasks.

Returns

A new Queue object

Method push()

Adds a task to the queue

Usage

Queue$push(fun, args = list(), id = NULL)

Arguments

fun

The function to be called when the task is scheduled

args

A list of arguments to be passed to the task function (optional)

id

A string specifying a unique identifier for the task (optional: tasks will be named "task_1", "task_2", etc if this is unspecified)

Returns

Invisibly returns the Task object

Method run()

Execute tasks in parallel using the worker pool, assigning tasks to workers in the same order in which they were added to the queue

Usage

Queue$run(
  timelimit = 60,
  message = "minimal",
  interval = 0.05,
  shutdown = TRUE
)

Arguments

timelimit

How long (in seconds) should the worker pool wait for a task to complete before terminating the child process and moving onto the next task? (default is 60 seconds, but this is fairly arbitrary)

message

What messages should be reported by the queue while it is running? Options are "none" (no messages), "minimal" (a spinner is shown alongside counts of waiting, running, and completed tasks), and "verbose" (in addition to the spinner, each task is summarized as it completes). Default is "minimal".

interval

How often should the task queue poll the workers to see if they have finished their assigned tasks? Specified in seconds.

shutdown

Should the workers in the pool be shut down (i.e., all R sessions closed) once the tasks are completed. Defaults to TRUE.

Returns

Returns a tibble containing the results of all executed tasks and various other useful metadata.

Method retrieve()

Retrieve the full state of the tasks queue in tidy form. If all tasks have completed this output is the same as the output as the run() method.

Usage

Queue$retrieve()

Returns

Returns a tibble containing the results of all executed tasks and various other useful metadata. Incomplete tasks nay have missing data.

Method get_queue_report()

Get as simplified description of the current state of the task queue. The corresponding private method is used internally to generate progress reports and to monitor progress of the task queue, but might be handy to have a public version.

Usage

Queue$get_queue_report()

Returns

Returns a tibble with three columns: task_id, state, and runtime

Method get_queue_workers()

Retrieve the workers

Usage

Queue$get_queue_workers()

Returns

A WorkerPool object

Method clone()

The objects of this class are cloneable with this method.

Usage

Queue$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.