R6 class storing a task list

R6 class storing a task list

R6 class storing a task list

Details

The TaskList class is used as a storage class. It provides a container that holds a collection of Task objects, along with a collection of methods for adding, removing, and getting Tasks. It can also report on the status of the Tasks contained within the list and retrieve results from those Tasks. What it cannot do is manage interactions with Workers or arrange for the Tasks to be executed. That's the job of the Queue.

Methods

Public methods

TaskList$new()

TaskList$length()

TaskList$add_task()

TaskList$remove_task()

TaskList$get_task()

TaskList$get_state()

TaskList$get_tasks_in_state()

TaskList$retrieve()

TaskList$clone()

Method new()

Create a new task list

Usage

TaskList$new()

Method length()

Return the number of tasks in the list

Usage

TaskList$length()

Returns

Integer

Method add_task()

Add a task to the TaskList

Usage

TaskList$add_task(task)

Arguments

task

The Task object to be added

Method remove_task()

This method removes one or more tasks from the TaskList.

Usage

TaskList$remove_task(x)

Arguments

x

Indices of the tasks to be removed

Method get_task()

Return a single Task contained in the TaskList. The Task is not removed from the TaskList, and has reference semantics: if the listed task is completed by a Worker, then the status of any Task returned by this method will update automatically

Usage

TaskList$get_task(x)

Arguments

x

The index the task to return

Returns

A Task object

Method get_state()

Return the status of all tasks in the TaskList. If requested, this method will also display messages summarising the current state of the tasks, and any tasks that have completed since the last time a status was returned. This messaging system is called by Queue objects as they work on a tasks

Usage

TaskList$get_state(message = "none", finished_in = NULL)

Arguments

message

Character specifying what type of message to display: "none" (the default), "minimal", or "verbose"

finished_in

A numeric value or a difftime specifying how long the tasks have taken to complete. This argument is only used when displaying messages, and it is used only to trigger the display of a tidy "all tasks completed" style message. It is purely cosmetic and does not affect the task status.

Returns

A character vector specifying the completion status for all listed tasks

Method get_tasks_in_state()

Return a list of tasks in a given state

Usage

TaskList$get_tasks_in_state(x)

Arguments

x

The name of the state (e.g., "waiting")

Returns

A TaskList object

Method retrieve()

Retrieve the full state of the tasks in tidy form. If all tasks have completed this output is the same as the output as the run() method for a Queue object.

Usage

TaskList$retrieve()

Returns

Returns a tibble containing the results of all executed tasks and various other useful metadata. Incomplete tasks may have missing data.

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskList$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.